backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
pydantic
pydantic-settings
python-dotenv
redis
aioredis
python-multipart
jose[cryptography]
passlib[bcrypt]
